requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
nodeenv==1.10.0
packaging==26.1
pluggy==1.6.0
prisma==0.15.0
pydantic==2.13.1
pydantic-settings==2.13.1
pydantic_core==2.46.1
Pygments==2.20.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-telegram-bot==22.7
starlette==1.0.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.34.0
